Extract projects
