+++ app/api/api/serializers.py
@@ -15,7 +15,6 @@ from shared.mapping.models import (
     MappingRule,
     OmopField,
     OmopTable,
-    Project,
     ScanReport,
     ScanReportConcept,
     ScanReportField,
@@ -734,37 +733,6 @@ class MappingRuleSerializer(DynamicFieldsMixin, serializers.ModelSerializer):
     class Meta:
         model = MappingRule
         fields = "__all__"
-
-
-class ProjectSerializer(DynamicFieldsMixin, serializers.ModelSerializer):
-    """
-    Serialiser for showing all details of a Project. Use in RetrieveViews
-    where User is permitted to view a particular Project.
-    """
-
-    class Meta:
-        model = Project
-        fields = "__all__"
-
-
-class ProjectNameSerializer(DynamicFieldsMixin, serializers.ModelSerializer):
-    """
-    Serialiser for only showing the names of Projects. Use in non-admin ListViews.
-    """
-
-    class Meta:
-        model = Project
-        fields = ["id", "name", "members"]
-
-
-class ProjectDatasetSerializer(DynamicFieldsMixin, serializers.ModelSerializer):
-    """
-    Serialiser for only showing the names of Projects. Use in non-admin ListViews.
-    """
-
-    class Meta:
-        model = Project
-        fields = ["name", "datasets", "members"]
 
 
 class GetRulesAnalysis(DynamicFieldsMixin, serializers.ModelSerializer):
