+++ app/api/projects/apps.py
@@ -1,0 +1,5 @@
+from django.apps import AppConfig
+
+
+class ProjectsConfig(AppConfig):
+    name = "projects"
